@@ -6,9 +6,7 @@ package io.flutter.plugins.geofencing
 
 import android.content.Context
 import android.content.Intent
-import android.os.IBinder
-import android.os.PowerManager
-import android.support.v4.app.JobIntentService
+import androidx.core.app.JobIntentService
 import android.util.Log
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
